Add Int to test size

+++ make.py
@@ -10,7 +10,7 @@ import shutil
 from argparser import args
 
 train_examples_per_shape = args.n
-test_examples_per_shape = max(1, train_examples_per_shape / 10)
+test_examples_per_shape = int(max(1, train_examples_per_shape / 10))
 
 image_w = 400
 image_h = 400
